Feature: added constructors for directed graph class
* Added constructors for directed graph based on adjacency lists.
* Added Planning.md
* Renamed adjacency list based graph class to "directed_graph".
* Removed KUNI::random namespace

The name change to directed_graph is for the sake of simplicity. This is
subject to change. The namespace (along with its files) was removed
because it was used for testing the meson build system. It served no
other purpose. Planning.md was added to place all plans for features or
ideas of implementing feature in one easily accessible place.